core: arm: plat-rz: add huk_read_root_material interface
Introduce huk.h and declare huk_read_root_material()
to eliminate missing prototype warnings for weak HUK
implementations.

+++ core/arch/arm/plat-rz/common/crypto/huk.c
@@ -10,6 +10,7 @@
 #include <mm/core_memprot.h>
 #include <tee_api_types.h>
 #include <tee/tee_cryp_pbkdf2.h>
+#include <huk.h>
 #include <platform_config.h>
 
 /*
@@ -20,12 +21,12 @@
  */
 __weak TEE_Result huk_read_root_material(uint8_t *buf, size_t *len)
 {
-	size_t i;
-	size_t read_len;
+	size_t	i;
+	size_t	read_len;
 	vaddr_t addr;
 
 	assert(buf && len);
-	read_len = MIN(*len, CHIPID_SIZE);
+	read_len = MIN(*len, (size_t)CHIPID_SIZE);
 
 	addr = (vaddr_t)phys_to_virt_io(CHIPID_BASE, CHIPID_SIZE);
 	assert(addr);
@@ -33,8 +34,7 @@ __weak TEE_Result huk_read_root_material(uint8_t *buf, size_t *len)
 	memset(buf, 0, *len);
 
 	for (i = 0; i < read_len; i += sizeof(uint32_t)) {
-		uint32_t value =
-			TEE_U32_TO_BIG_ENDIAN(io_read32(addr + i));
+		uint32_t value = TEE_U32_TO_BIG_ENDIAN(io_read32(addr + i));
 
 		memcpy(buf + i, &value, MIN(read_len - i, sizeof(value)));
 	}
@@ -61,18 +61,22 @@ __weak TEE_Result huk_read_root_material(uint8_t *buf, size_t *len)
 static TEE_Result huk_kdf(uint8_t *huk, size_t huk_length)
 {
 	uint8_t password[HW_UNIQUE_KEY_LENGTH];
-	size_t password_len = sizeof(password);
+	size_t	password_len = sizeof(password);
 
 	static const uint32_t huk_pbkdf2_iteration_count = 1000U;
 
-	static const uint8_t huk_pbkdf2_salt[] = {0x76, 0x6A, 0xEF, 0x5C, 0x39, 0xEF, 0x6C, 0x26, 0x41, 0x6C, 0x46, 0x68, 0x05, 0x43, 0x06, 0xC0};
+	static const uint8_t huk_pbkdf2_salt[] = { 0x76, 0x6A, 0xEF, 0x5C,
+						   0x39, 0xEF, 0x6C, 0x26,
+						   0x41, 0x6C, 0x46, 0x68,
+						   0x05, 0x43, 0x06, 0xC0 };
 
 	TEE_Result res = huk_read_root_material(password, &password_len);
 	if (res != TEE_SUCCESS)
 		return res;
 
-	return tee_cryp_pbkdf2(TEE_ALG_HMAC_SHA256, password, password_len, huk_pbkdf2_salt, sizeof(huk_pbkdf2_salt), huk_pbkdf2_iteration_count,
-				huk, huk_length);
+	return tee_cryp_pbkdf2(TEE_ALG_HMAC_SHA256, password, password_len,
+			       huk_pbkdf2_salt, sizeof(huk_pbkdf2_salt),
+			       huk_pbkdf2_iteration_count, huk, huk_length);
 }
 
 TEE_Result tee_otp_get_hw_unique_key(struct tee_hw_unique_key *hwkey)
